Update Script
For the most part it runs fine. I changed the sql query in line 90 to match what the sql query is looking for as far as format. I also changed line 149 to take in the columns of the dataframe. One thing that is not functional is line 128. It is throwing an error because it is taking in a string in the column name and trying to make it a float. "ValueError: could not convert string to float: 'EYERHALL_03_RM_4427_DAYLIGHT'". Also we need to find a way to change the format of the data from 'mm/dd/yyyy' to 'mm_dd_yyyy' because you cannot have a file name with a '/'. But that is the format needed for sql query. Also we should setup some possible values for frequency.